Move 'use auto prop' into code style layer
