+++ src/Workspaces/Core/Portable/FindSymbols/FindReferences/Finders/AbstractReferenceFinder.cs
@@ -600,156 +600,8 @@ internal abstract partial class AbstractReferenceFinder : IReferenceFinder
         FindReferencesDocumentState state,
         CancellationToken cancellationToken)
     {
-        var syntaxFacts = state.SyntaxFacts;
-        var semanticFacts = state.SemanticFacts;
-        var semanticModel = state.SemanticModel;
-
-        var topNameNode = node;
-        while (syntaxFacts.IsQualifiedName(topNameNode.Parent))
-            topNameNode = topNameNode.Parent;
-
-        var parent = topNameNode?.Parent;
-
-        // typeof/sizeof are a special case where we don't want to return a TypeOrNamespaceUsageInfo, but rather a ValueUsageInfo.Name.
-        // This brings it in line with nameof(...), making all those operators appear in a similar fashion.
-        if (parent?.RawKind == syntaxFacts.SyntaxKinds.TypeOfExpression ||
-            parent?.RawKind == syntaxFacts.SyntaxKinds.SizeOfExpression)
-        {
-            return new(ValueUsageInfo.Name, typeOrNamespaceUsageInfoOpt: null);
-        }
-
-        var isInNamespaceNameContext = syntaxFacts.IsBaseNamespaceDeclaration(parent);
-        return syntaxFacts.IsInNamespaceOrTypeContext(topNameNode)
-            ? SymbolUsageInfo.Create(GetTypeOrNamespaceUsageInfo())
-            : GetSymbolUsageInfoCommon();
-
-        // Local functions.
-        TypeOrNamespaceUsageInfo GetTypeOrNamespaceUsageInfo()
-        {
-            var usageInfo = IsNodeOrAnyAncestorLeftSideOfDot(node, syntaxFacts) || syntaxFacts.IsLeftSideOfExplicitInterfaceSpecifier(node)
-                ? TypeOrNamespaceUsageInfo.Qualified
-                : TypeOrNamespaceUsageInfo.None;
-
-            if (isInNamespaceNameContext)
-            {
-                usageInfo |= TypeOrNamespaceUsageInfo.NamespaceDeclaration;
-            }
-            else if (node.FirstAncestorOrSelf<SyntaxNode, ISyntaxFactsService>((node, syntaxFacts) => syntaxFacts.IsUsingOrExternOrImport(node), syntaxFacts) != null)
-            {
-                usageInfo |= TypeOrNamespaceUsageInfo.Import;
-            }
-
-            while (syntaxFacts.IsQualifiedName(node.Parent))
-                node = node.Parent;
-
-            if (syntaxFacts.IsTypeArgumentList(node.Parent))
-            {
-                usageInfo |= TypeOrNamespaceUsageInfo.TypeArgument;
-            }
-            else if (syntaxFacts.IsTypeConstraint(node.Parent))
-            {
-                usageInfo |= TypeOrNamespaceUsageInfo.TypeConstraint;
-            }
-            else if (syntaxFacts.IsBaseTypeList(node.Parent) ||
-                syntaxFacts.IsBaseTypeList(node.Parent?.Parent))
-            {
-                usageInfo |= TypeOrNamespaceUsageInfo.Base;
-            }
-            else if (syntaxFacts.IsTypeOfObjectCreationExpression(node))
-            {
-                usageInfo |= TypeOrNamespaceUsageInfo.ObjectCreation;
-            }
-
-            return usageInfo;
-        }
-
-        SymbolUsageInfo GetSymbolUsageInfoCommon()
-        {
-            if (semanticFacts.IsInOutContext(semanticModel, node, cancellationToken))
-            {
-                return SymbolUsageInfo.Create(ValueUsageInfo.WritableReference);
-            }
-            else if (semanticFacts.IsInRefContext(semanticModel, node, cancellationToken))
-            {
-                return SymbolUsageInfo.Create(ValueUsageInfo.ReadableWritableReference);
-            }
-            else if (semanticFacts.IsInInContext(semanticModel, node, cancellationToken))
-            {
-                return SymbolUsageInfo.Create(ValueUsageInfo.ReadableReference);
-            }
-            else if (semanticFacts.IsOnlyWrittenTo(semanticModel, node, cancellationToken))
-            {
-                return SymbolUsageInfo.Create(ValueUsageInfo.Write);
-            }
-            else
-            {
-                var operation = semanticModel.GetOperation(node, cancellationToken);
-                if (operation is IObjectCreationOperation)
-                    return SymbolUsageInfo.Create(TypeOrNamespaceUsageInfo.ObjectCreation);
-
-                // Note: sizeof/typeof also return 'name', but are handled above in GetSymbolUsageInfo.
-                if (operation?.Parent is INameOfOperation)
-                    return SymbolUsageInfo.Create(ValueUsageInfo.Name);
-
-                if (node.IsPartOfStructuredTrivia())
-                    return SymbolUsageInfo.Create(ValueUsageInfo.Name);
-
-                var symbolInfo = semanticModel.GetSymbolInfo(node, cancellationToken);
-                if (symbolInfo.Symbol != null)
-                {
-                    switch (symbolInfo.Symbol.Kind)
-                    {
-                        case SymbolKind.Namespace:
-                            var namespaceUsageInfo = TypeOrNamespaceUsageInfo.None;
-                            if (isInNamespaceNameContext)
-                                namespaceUsageInfo |= TypeOrNamespaceUsageInfo.NamespaceDeclaration;
-
-                            if (IsNodeOrAnyAncestorLeftSideOfDot(node, syntaxFacts))
-                                namespaceUsageInfo |= TypeOrNamespaceUsageInfo.Qualified;
-
-                            return SymbolUsageInfo.Create(namespaceUsageInfo);
-
-                        case SymbolKind.NamedType:
-                            var typeUsageInfo = TypeOrNamespaceUsageInfo.None;
-                            if (IsNodeOrAnyAncestorLeftSideOfDot(node, syntaxFacts))
-                                typeUsageInfo |= TypeOrNamespaceUsageInfo.Qualified;
-
-                            return SymbolUsageInfo.Create(typeUsageInfo);
-
-                        case SymbolKind.Method:
-                        case SymbolKind.Property:
-                        case SymbolKind.Field:
-                        case SymbolKind.Event:
-                        case SymbolKind.Parameter:
-                        case SymbolKind.Local:
-                            var valueUsageInfo = ValueUsageInfo.Read;
-                            if (semanticFacts.IsWrittenTo(semanticModel, node, cancellationToken))
-                                valueUsageInfo |= ValueUsageInfo.Write;
-
-                            return SymbolUsageInfo.Create(valueUsageInfo);
-                    }
-                }
-
-                return SymbolUsageInfo.None;
-            }
-        }
-    }
-
-    private static bool IsNodeOrAnyAncestorLeftSideOfDot(SyntaxNode node, ISyntaxFactsService syntaxFacts)
-    {
-        if (syntaxFacts.IsLeftSideOfDot(node))
-        {
-            return true;
-        }
-
-        if (syntaxFacts.IsRightOfQualifiedName(node) ||
-            syntaxFacts.IsNameOfSimpleMemberAccessExpression(node) ||
-            syntaxFacts.IsNameOfMemberBindingExpression(node))
-        {
-            return syntaxFacts.IsLeftSideOfDot(node.Parent);
-        }
-
-        return false;
+        return SymbolUsageInfo.GetSymbolUsageInfo(
+            state.SemanticFacts, state.SemanticModel, node, cancellationToken);
     }
 
     internal static ImmutableArray<(string key, string value)> GetAdditionalFindUsagesProperties(
